Refactor storage and associated tests
