+++ src/main/java/dash/MainApp.java
@@ -21,14 +21,14 @@ import dash.model.ReadOnlyUserPrefs;
 import dash.model.UserPrefs;
 import dash.model.task.TaskList;
 import dash.model.util.SampleDataUtil;
-import dash.storage.AddressBookStorage;
-import dash.storage.JsonAddressBookStorage;
-import dash.storage.JsonUserPrefsStorage;
 import dash.storage.Storage;
 import dash.storage.StorageManager;
-import dash.storage.UserPrefsStorage;
+import dash.storage.addressbook.AddressBookStorage;
+import dash.storage.addressbook.JsonAddressBookStorage;
 import dash.storage.tasklist.JsonTaskListStorage;
 import dash.storage.tasklist.TaskListStorage;
+import dash.storage.userprefs.JsonUserPrefsStorage;
+import dash.storage.userprefs.UserPrefsStorage;
 import dash.ui.Ui;
 import dash.ui.UiManager;
 import javafx.application.Application;
