+++ src/main/java/dash/storage/StorageManager.java
@@ -11,7 +11,9 @@ import dash.model.ReadOnlyAddressBook;
 import dash.model.ReadOnlyUserPrefs;
 import dash.model.UserPrefs;
 import dash.model.task.TaskList;
+import dash.storage.addressbook.AddressBookStorage;
 import dash.storage.tasklist.TaskListStorage;
+import dash.storage.userprefs.UserPrefsStorage;
 
 /**
  * Manages storage of AddressBook data in local storage.
