+++ src/test/java/dash/logic/LogicManagerTest.java
@@ -23,10 +23,10 @@ import dash.model.ReadOnlyAddressBook;
 import dash.model.UserPrefs;
 import dash.model.person.Person;
 import dash.model.task.TaskList;
-import dash.storage.JsonAddressBookStorage;
-import dash.storage.JsonUserPrefsStorage;
 import dash.storage.StorageManager;
+import dash.storage.addressbook.JsonAddressBookStorage;
 import dash.storage.tasklist.JsonTaskListStorage;
+import dash.storage.userprefs.JsonUserPrefsStorage;
 import dash.testutil.Assert;
 import dash.testutil.PersonBuilder;
 import dash.testutil.TypicalPersons;
