+++ src/test/java/dash/storage/addressbook/JsonAddressBookStorageTest.java
@@ -1,4 +1,4 @@
-package dash.storage;
+package dash.storage.addressbook;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
